+++ src/types/materials/index.ts
@@ -3,6 +3,7 @@ export * from './material-create';
 export * from './material-update';
 export * from './material-location-threshold';
 export * from './low-stock-material';
+export * from './material-stock-summary';
 export * from './material-consumption';
 export * from './material-consumption-create';
 export * from './enum';